renamed interface.py to smrf_connector, cleaned up imports
