--- awsm/interface/__init__.py
@@ -1,7 +1,0 @@
-# -*- coding: utf-8 -*-
-
-__author__ = """Micah Sandusky"""
-__email__ = 'micah.sandusky@ars.usda.gov'
-
-from . import interface, ipysnobal, smrf_ipysnobal, \
-    pysnobal_io, ingest_data
